pyproject.toml
[tool.poetry]
name = "sudokus"
version = "0.1.0"
description = ""
authors = ["mattjohnrock <johnstone.mattjames@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
